:mag: update metadata
