+++ app/layout.tsx
@@ -2,12 +2,6 @@ import '@/styles/globals.css';
 
 import React from 'react';
 import { Footer } from '@/ui/Footer';
-import { Metadata } from 'next';
-import { Meta } from './meta';
-
-export const metadata: Metadata = {
-  description: 'ブレスオブザワイルドの自撮りハイラル図鑑',
-};
 
 export default function RootLayout({
   children,
@@ -16,7 +10,6 @@ export default function RootLayout({
 }) {
   return (
     <html>
-      <Meta />
       <body className="overflow-y-scroll  bg-zinc-900">
         <div className="py-8 sm:mx-1 lg:mx-auto lg:max-w-6xl">
           <div className="rounded-xl border border-zinc-800 bg-black lg:px-8 lg:py-8 md:px-1 md:py-6">
